dev-requirements.txt
WebTest
nose
nose-cov
pyramid_debugtoolbar
waitress
